Add batch size to command.
