+++ src/indexables/application/next-verification-action-handler.php
@@ -8,6 +8,7 @@ use Yoast\WP\SEO\Indexables\Domain\Current_Verification_Action;
 use Yoast\WP\SEO\Indexables\Domain\Last_Batch_Count;
 
 class Next_Verification_Action_Handler {
+
 	/**
 	 * @var Options_Helper
 	 */
@@ -21,7 +22,7 @@ class Next_Verification_Action_Handler {
 		return $this->options_helper->get( 'cron_verify_current_action', 'term' );
 	}
 
-	public function set_current_verification_action( Current_Verification_Action $current_verification_action) {
+	public function set_current_verification_action( Current_Verification_Action $current_verification_action ) {
 		$this->options_helper->set( 'cron_verify_current_action', $current_verification_action->get_action() );
 	}
 }
